The class GecodeSpace divided into gecode-space-decl.hh and gecode-space.hh
